FormulasProposicionales

Description

Práctica final de Programación Declarativa (PD) de Ingeniería Informática de la UCM del curso 2022-2023

Types

data Prop

Atomic predicate

data Formula

Formula constructor

type Clausula = [Prop]

Alias for list os propositions as clause

Functions

Parsers

fncAlista :: Formula -> [Clausula]

Description

Translates a CNF function to a list of clauses

Example

Expand

>>> fncAlista (And (Or (Atom "A") (Atom "B")) (Not (Atom "B")))
[[Pos "A",Pos "B"],[Neg "B"]]

clausulaLista :: Formula -> Clausula

Description

Translates a Formula wich without (^) to a list of proposition

Example

Expand

>>> clausulaLista (Not (Atom "q"))
[Neg "q"]

>>> clausulaLista (Or (Not (Atom "Q")) (Atom "R"))
[Neg "Q",Pos "R"]

formulaAfnc :: Formula -> Formula

Description Applies distributiva and deMorgan to a Formula to change it to its NCF

Example

Expand

>>> formulaAfnc (And (Not (And (Atom "A") (Atom "B"))) (Atom "C"))
((!"A"^"C")v(!"B"^"C"))

listaAtupla :: [Clausula] -> [Clausula] -> [(Clausula, Clausula)]

Description

Aux function for resolucion that adds two list of lists to a list of tuples of each list

Examples

Expand

>>> listaAtupla [[Neg "A"], [Pos "A", Neg "B"]] [[Neg "A"], []]
[([Neg "A"],[Neg "A"]),([Pos "A",Neg "B"],[])]

Utilities

esPositivo :: Prop -> Bool

Description

Checks if a proposition is positive or negative

Example

Expand

>>> esPositivo (Pos "q")
True

>>> esPositivo (Neg "q")
False

esClausula :: Formula -> Bool

Description

Checks if a formula can be converted to a clause

Example

Expand

>>> esClausula (Or (Atom "A") (And (Atom "B") (Atom "C")))
False

>>> esClausula (Or (Not (Atom "A")) (Atom "B"))
True

borrarLista :: [Clausula] -> Clausula -> [Clausula]

Description Aux function for resolucion that erases a Clause that is no longer needed

Example

Expand

>>> borrarLista [[Neg "Q", Neg "S"], [Pos "S", Neg "Q"]] [Pos "S", Neg "Q"]
[[Neg "Q",Neg "S"]]

indice :: [Clausula] -> Clausula -> Int -> Int

Description

Aux function for resolucion that gets the index of a given clause

Example

Expand

>>> indice [[Neg "Q", Neg "S"], [Pos "S", Neg "Q"]] [Pos "S", Neg "Q"] 0
1

Operations on Formulas

esClausulaHorn :: Clausula -> Bool

Description

Checks if a given clause is a Horn's clause (one or zero positive propositions)

Example

Expand

>>> esClausulaHorn [Neg "A", Pos "B", Pos "C"]
False

>>> esClausulaHorn [Neg "A"]
True

>>> esClausulaHorn [Neg "A", Neg "B", Pos "C"]
True

resolvente :: Clausula -> Clausula -> Clausula

Description

Given a Clausula gives the resolvents with another Clausula

Example

Expand

>>> resolvente (clausulaLista (Or (Atom "P") (Or (Not (Atom "Q")) (Atom "R")))) (clausulaLista (Or (Atom "P") (Or (Not (Atom "R")) (Atom "S"))))
[Neg "Q",Pos "P",Pos "S"]

deMorgan :: Formula -> Formula

Description

Applies deMorgan properties to a Formula

Example

Expand

>>> deMorgan (Not (And (Atom "A") (Atom "B")))
(!"A"v!"B")

>>> deMorgan (Not (Or (Atom "A") (Atom "B")))
(!"A"^!"B")

>>> deMorgan (Not (Or (Atom "A") (Not (Atom "B"))))
(!"A"^"B")

>>> deMorgan (Not (And ((Not (Or (Atom "A") (Atom "B")))) (Atom "B")))
(("A"v"B")v!"B")

distributiva :: Formula -> Formula

Description

Applies distributed properties to a Formula but only to transform to NCF, not complete distributed.

Example

Expand

>>> distributiva (And ((Or (Atom "A") (Atom "B"))) (Atom "B"))
(("A"^"B")v("B"^"B"))

>>> distributiva (Not (Or (And (Atom "A") (Atom "B")) (Atom "C")))
!(("A"v"C")^("B"v"C"))

resolucion :: [Clausula] -> Clausula -> IO Clausula

Descripcion

Makes resolvants for a function until no more resolvants are posible

IO

getFormula :: IO Formula

Description

Reads a Formula from the standard input

getInt :: IO Int

Reads an int from the standard input

menu :: Formula -> Formula -> IO ()

Description

Simple menu to test functions

demoMode :: IO ()

Description

Demo for the program

Try: f: (And (And (Or(Atom A)(Not (Atom B)))(Or (Not (Atom A)) (Atom B))) (Atom A)) r: (Not (Atom B))