Interprete

Description

Práctica final de Programación Declarativa (PD) de Ingeniería Informática de la UCM del curso 2023-2024

DATA TYPES

type Var = String

Alias for Variables

type Program = [Instruction]

Alias for a list of instructions

showProgram :: Program -> String

Description

Shows a full program

type VarVal = (Var, Int)

Alias for the program way of handling with variables

type State = [VarVal]

Alias for the program stack

mostrarStack :: State -> IO ()

Description

Shows the current state of the stack and waits for an enter key press to continue execution

data Symbol

Posible arithmetic operations in the program

Constructors

Plus

Minus

Times

Equals

data Comparator

Posible logic operations in the program

Constructors

Greater

Less

GEQ

LEQ

Eq

Not

And

Or

data Operation

Arithmetic operations in the program. Can use variables or integers.

Constructors

O
Full operation

Fields

Operation

variable to change value

Symbol

arithmetic operation

Operation

Operation to perform

I
Operation is only an integer

Fields

Int

Integer to operate with

V
Operation uses a variable

Fields

Var

Variable to get the value from

data Instruction

Instructions inside the program, can be assigning new values to variables or while loops

Constructors

A
Assign operations

Fields

Var

Variable to change value

Operation

Operation to change value

W
While loop

Fields

Comp

Comparator to end the loop

Program

Set of instructions inside the loop

data Comp

Boolean comparators for while loops (W)

Constructors

CO
Arithmetic comparator

Fields

Operation

Variable to compare

Comparator

Comparator used

Operation

Integer to compare to

CB
Boolean Comparator

Fields

Bool

First part of the comparison

Comparator

Comparator used

Bool

Second part of the comparison

CC
Comparator of comparators

Fields

Comp

Variable to compare

Comparator

Comparator used

Comp

Integer to compare to

Void
Void comparator, used for not clauses

STACK OPERATIONS

isVal

Arguments

:: State
current stack of the program

-> Var
Variable to check

-> Bool
Variable is in stack or not

Description

Checks if a given variable name is in the stack.

getVal

Arguments

:: State
Current stack of the program

-> Var
Variable to check

-> Int
Value of the given variable

Description

Gives a the stored value for a value for a variable. If the variable is not in the stack, the function throws an error.

setVal

Arguments

:: State
Current stack of the program

-> Var
Variable to update

-> Int
New value

-> State
Updated stack

Description

Changes the value of a given variable to a new one.

INSTRUCTION OPERATIONS

resolver :: Comp -> State -> Bool

Description

Computes a boolean from the different comparator types

calc

Arguments

:: Int
First int

-> Symbol
Operation to perform

-> Int
Second int

-> Int
Result

Description

Performs an arithmetic operation (Symbol) with two Integers

calcular

Arguments

:: Operation
Operation to perform

-> State
Current state of the program

-> Int
Result of the operation

Description

Performs an arithmetic operation given the operation and the current stack of the program

EXECUTION

ejecutarAux

Arguments

:: Instruction
Instruction to execute

-> State
Current state of the program

-> State
Next state of the program

Description

Auxiliary function for ejecutar splitted into the different types of instructions available.

ejecutar

Arguments

:: Program
Set of instructions to compute

-> State
Current state of the program

-> State
New state of the program

Description

Executes all the instructions in a program in sequence

ejecuta

Arguments

:: Program

-> State
Initial state

-> Int
Value of the last initialized variable

Description

Executes all the instructions in a given program with a given initial state and returns the last variable declared

IO EXECUTION

ejecutarAuxIO

Arguments

:: Instruction

-> State
Current state of the program

-> IO State
Next state of the program

Description

Executes an instruction showing the current state of the stack

ejecutarIO

Arguments

:: Program
Set of instructions

-> State
Initial stack

-> IO State
Value of last initialized variable

Description

Executes an entire program showing the state of the stack before every instruction. See ejecutarAuxIO

ejecutaIO

Arguments

:: Program
Set of instructions

-> State
Initial stack

-> IO Int
Value of last initialized variable

Description

Executes an entire program and returns the value of the R variable showing the state of the stack before every instruction. See ejecutarIO.

EXAMPLES

s0 :: State

Initial state for factorial

s1 :: State

Initial state for factorial

s2 :: State

Initial state for factorial

factorial :: Program

Description

Factorial calculation for a number given

 Y := X
 R := 1
 while (Y > 0) {
  R := R :* Y
  Y := Y :-1
 }

s3 :: State

Initial state for sumador

s4 :: State

Initial state for sumador

s5 :: State

initial state for sumador

sumador :: Program

Description

Summation of from 1 to a given number

i := 1
R := 0
while (i < X && R != INT_MAX){
  R := i :+ R
  i := i :+1
}

EXECUTION MODES

testResults

Arguments

:: Bool
debug mode enabled

-> Program
Instructions to execute

-> State
Initial state for the program

-> Int
Expected result of the program

-> IO ()

Description

Executes a program with a given state and shows the result compared to the expected result

testMode

Arguments

:: Bool
debug mode enabled

-> IO ()

Description

Execution of varios programs with different initial stacks and showing results compared to expected results

freeMode

Arguments

:: Bool
debug mode enabled

-> IO ()

Description

Lets the user write their own program and initial state and shows the result

fileMode

Arguments

:: Bool
Debug mode enabled

-> IO ()

Description

Lets the user give a file with a written program an then asks for a initial state. Shows results after execution

UTILITIES

getInt :: IO Int

Descripción

Coge un número desde la entrada.

entrypoint :: IO ()

Description

Main menu for the program, lets the user choose mode and graphical options

main :: IO ()

Description

Entrypoint