Labo1

Description

EJERCICIO 1

A)

time :: Integer

tiempo base del ejercicio 1

yearsSeg :: Integer

Segundos totales en un año

daysSeg :: Integer

Segundos totales en un día

hoursSeg :: Integer

Segundos totales en una hora

minsSeg :: Integer

Segundos totales en un minuto

years :: Integer

Caclulo de los años equivalentes a time

days :: Integer

Calculo de las horas equivalentes a time quitando years

hours :: Integer

Calculo de las horas equivalentes a time quitando days y years

mins :: Integer

Calculo de las horas equivalentes a time quitando days, years y hours

segs :: Integer

Calculo de las horas equivalentes a time quitando days, years, hours y mins

sol1a :: (Integer, Integer, Integer, Integer, Integer)

Tupla con el tiempo total de time en años, dias, horas, minutos y segundos

B)

sol1b :: Integer -> (Integer, Integer, Integer, Integer, Integer)

Descripción

Coje el tiempo dado y lo convierte a (Años, Días, Horas, Minutos, Segundos)

Ejemplos

>>> sol1b 1000000
(0,11,13,46,40)

EJERCICIO 2

A)

sol2A :: Integer -> Bool

Descripción

Coje el año dado y devuelve True si es bisiesto, False en caso contrario

Ejemplos

>>> sol2A 2008
True

>>> sol2A 2022
False

B)

sol2B :: Integer -> Bool

Descripción

Coje el año dado y devuelve True si es bisiesto, False en caso contrario

Ejemplos

>>> sol2A 2008
True

>>> sol2A 2022
False

EJERCICIO 3

sol3 :: (Fractional a1, Integral a2, Foldable t) => t a2 -> a1

Descripción

Error de tipado ya que la division / no trabaja con Integer

EJERCICIO 4

calculoDigitos :: Integer -> Integer

Descripción

Cálculo del número de dijitos de un numero dado

Ejemplos

>>> calculoDigitos 0
1

>>> calculoDigitos 1984
4

sumaDigitos :: Integral t => t -> t

Descripción

Suma todos los digitos de un numero dado

Ejemplos

>>> sumaDigitos 0
0

>>> sumaDigitos 1984
22

reduccion :: Integral t => t -> t

Descripción

Suma todos los digitos de un numero dado hasta que el total es menor que 10

Ejemplos

>>> reduccion 0
0

>>> reduccion 1984
4

EJERCICIO 5

bools :: Bool -> Bool -> Bool

Disyunción booleana definida por ajuste de patrones