Labo3

Description

EJERCICIO 1

A)

last' :: [a] -> a

Descripción

Coge el último elemento de una lista dada

Ejemplos

>>> last' []
Lista vacia

>>> last [1,2,3,4]
4

B)

reverse' :: [a] -> [a]

Descripción

Hace la lista inversa de una dada

Ejemplos

>>> reverse' [1,2,3,4]
[4,3,2,1]

C)

all' :: (a -> Bool) -> [a] -> Bool

Descripción

Devuelve true si todos los elementos de la lista cumplen una condición

Ejemplos

>>> all' (<10) [1,2,3,4]
True

>>> all' (>10) [1,20,30,40]
False

D)

min' :: Ord a => [a] -> a

Descripción

Devuelve el menor elemento de una lista dada

Ejemplos

>>> min' []
Lista vacia

>>> min' [2,5,8,2,0]
0

E)

map' :: (a -> b) -> [a] -> [b]

Descripción

Genera una lista de elementos resultantes de aplicar un filtro a otra lista de elementos

Ejemplo

>>> map' (*3) [1,2,3,4]
[3,6,9,12]

F)

filter' :: (a -> Bool) -> [a] -> [a]

Descripción

Genera una lista de elementos con los elementos de otra lista que hayan pasado un filtro

Ejemplo

>>> filter' (>10) [1,20,30,40]
[20,30,40]

G)

takeWhile' :: (a -> Bool) -> [a] -> [a]

Descripción

Coge elementos de la lista dada hasta que la condición falla

Ejemplos

>>> takeWhile' (<3) [1,2,3,4,5]
[1,2]

>>> takeWhile' (>3) [1,2,3,4,5]
[]

EJERCICIO 2

listaNegPos :: [Integer]

Descripción

Genera una lista alternada de elementos del 1 al 100 de forma [1, -1]

EJERCICIO 3

listaParejas' :: (Num b, Enum b, Eq b) => b -> [(b, b)]

Descripción

Versión acotada de listaParejas

Ejemplo

>>> listaParejas' 4
[(0,0),(0,1),(1,0),(0,2),(1,1),(2,0),(0,3),(1,2),(2,1),(3,0),(0,4),(1,3),(2,2),(3,1),(4,0)]

listaParejas :: Integral a => [(a, a)]

Descripción

Genera parejas ordenadas que cumplen x + y = z, en orden de z

Ver listaParejas' para ejemplos

EJERCICIO 4

A)

sufijos :: [a] -> [[a]]

Descripción

Genera todos los sufijos de una lista

Ejemplo

>>> sufijos [1,2,3,4]
[[1,2,3,4],[2,3,4],[3,4],[4],[]]

B)

sublistas :: [a] -> [[a]]

Descripción

Genera todas las sublistas de una lista dada

Ejemplo

>>> sublistas [1,2,3]
[[1],[2],[3],[1,2],[2,3],[1,2,3]]

C)

permuta :: [a] -> [[a]]

Descripción

Genera todas las permutaciones posibles de una lista dada

Ejemplo

>>> permuta [1,2,3]
[[1,2,3],[2,1,3],[2,3,1],[1,3,2],[3,1,2],[3,2,1]]

intercalado :: a -> [a] -> [[a]]

Descripción

Devuelve una lista de listas con el elemento dado en cada posición posible de la lista dada

Ejemplo

>>> intercalado 3 [1,2]
[[3,1,2],[1,3,2],[1,2,3]]

D)

sumandos :: Integral a => a -> [[a]]

Descripción

Genera un listado de de listas que al sumar todos sus elementos dan como resultado el numero dado

Ejemplos

>>> sumandos 3
[[1,1,1],[1,2],[3]]

sumandos' :: Integral a => a -> a -> a -> [[a]]

Descripción

Genera listados de tamaños desde x hasta n que al ser sumados dan como resultado x, siendo i el numero de listados generados

Ejemplo

>>> sumandos' 4 1 1
[[1,1,1],[1,2],[3]]

>>> sumandos' 4 2 1
[[1,1],[2]]

>>> sumandos' 4 2 2
[[2]]