Labo4

Description

EJERCICIO 1

data Punto

Representación de un punto de forma (x,y)

Constructors

P
Constructor del punto

Fields

Int

Coordenada X

Int

Coordenada Y

pointSum :: (Ord a, Num a) => a -> a -> a

Descripción

Suma un numero a una parte de una coordenada dada. Tiene como rango [0 .. 100]

Ejemplos

>>> pointSum 100 10
9

>>> pointSum 0 (-1)
100

>>> pointSum 100 1
0

data Direccion

Representación de una dirección

Constructors

ARRIBA

ABAJO

IZQUIERDA

DERECHA

A)

mueve :: Punto -> Direccion -> Punto

Descripción

Mueve un punto en la dirección dada

Ejemplos

>>> mueve (P 0 1) IZQUIERDA
(100,1)

>>> mueve (P 0 1) DERECHA
(1,1)

>>> mueve (P 0 1) ARRIBA
(0,2)

>>> mueve (P 0 1) ABAJO
(0,0)

B)

destino :: Punto -> [Direccion] -> Punto

Descripción

Mueve un punto en las direcciones dadas

Ejemplos

>>> destino (P 0 0) [ARRIBA, ARRIBA, ABAJO, ABAJO, IZQUIERDA, DERECHA, IZQUIERDA, DERECHA]
(0,0)

>>> destino (P 0 0) [ARRIBA, ARRIBA, ABAJO, IZQUIERDA, IZQUIERDA, DERECHA]
(100,1)

C)

trayectoria :: Punto -> [Direccion] -> [Punto]

Descripción

Mueve un punto en las direcciones dadas mostrando los puntos por los que se han pasado

Ejemplos

>>> trayectoria (P 0 0) [ARRIBA, ARRIBA, ABAJO, ABAJO, IZQUIERDA, DERECHA, IZQUIERDA, DERECHA]
[(0,0),(0,1),(0,2),(0,1),(0,0),(100,0),(0,0),(100,0),(0,0)]

>>> trayectoria (P 0 0) [ARRIBA, ARRIBA, ABAJO, IZQUIERDA, IZQUIERDA, DERECHA]
[(0,0),(0,1),(0,2),(0,1),(100,1),(99,1),(100,1)]

EJERCICIO 2

data Nat

Representación de numeros naturales según la aritmética de Peano

Constructors

Cero
Representación del 0

Suc Nat
Representación de cualquier sucesor de 0

A)

(~+) :: Nat -> Nat -> Nat infix 4

Descripción

Suma de dos números siguiendo la aritmética de Peano

Ejemplo

>>> (Suc (Cero)) ~+ (Suc (Suc (Cero)))
3

(~*) :: Nat -> Nat -> Nat infix 6

Descripción

Multiplicación de dos números siguiendo la aritmética de Peano

Ejemplo

>>> (Suc (Cero)) ~* (Suc (Suc (Cero)))
2

B)

natToInt :: Nat -> Int

Descripción

Convierte un número representado en aritmética de Peano en un Int normal

Ejemplo

>>> natToInt (Suc (Suc (Cero)))
2

C)

EJERCICIO 3

data Complejo

Representación de un numero complejo

Constructors

C Float Float

EJERCICIO 4

class Medible a where

Clase para medir otros tipos de datos

Methods

medida

Arguments

:: a

-> Int
Coge el dato dado y saca una medida de él