Labo5

Description

EJERCICIO 1

getInt :: IO Int

Descripción

Coge un número desde la entrada.

adivina :: Int -> IO ()

Descripión

Pide un número por teclado a adivinar. Esto es bastante estúpido de por si solo porque tienes que introducir el número como argumento

EJERCICIO 2

type Matriz = [[Float]]

Representación de una matriz como lista de listas

getFloat :: IO Float

Descripción

Coge un número desde la entrada.

A)

getMatriz :: IO Matriz

Descripción

Coge una matriz desde la entrada.

Ver getDatosMatriz para ver el funcionamiento total

getFilas :: Int -> IO [Float]

Descripción

Coge una fila de una matriz según el número de columnas dadas.

getDatosMatriz :: Int -> Int -> IO Matriz

Descripción

Coge todas las filas de una matriz según el número de filas y columnas dadas.

Ver getFilas para el funcionamiento completo

B)

dibujaMatriz :: Matriz -> IO ()

Descripción

Dibuja una matriz dada como argumento.

mostrarMatrizNueva :: IO ()

Descripción

Pide una matriz por teclado y la muestra una vez construida.

Ver getMatriz y dibujaMatriz.

EJERCICIO 3

formatea :: String -> String -> Int -> IO ()

Descripción

Formatea a n columnas cada linea de un fichero dado en otro.

Ver formatear para el funcionamiento.

formatear :: Int -> String -> String

Descripción

Formatea cada linea a un total de n caracteres por fila.

Ver justify para el funcionamiento completo.

justify :: Int -> String -> String

Descripción

Añade espacios a una línea si es necesario.

Ver addSpaces para el funcionamiento completo.

addSpaces :: Int -> String -> String

Descripción

Añade n espacios a una lina dada.