Lote1

Description

EJERCICIO 1

A)

petaFlop :: Integer

Flops in a petaFlop

fornl :: Integer

Computing power of the best computer of Frontier

yearsToSec :: Integer

Seconds in a year

frontierInteger :: Integer

Power of operation of the Frontier since the beginning of time | >>> frontierInteger 515859580800000000000000000000000000

petaFlop' :: Int

Flops in a petaFlop

fornl' :: Int

Computing power of the best computer of Frontier

yearsToSec' :: Int

Seconds in a year

frontierInt :: Int

Power of operation of the Frontier since the beginning of time | >>> frontierInt -7928755680649936896

petaFlop'' :: Double

Flops in a petaFlop

fornl'' :: Double

Computing power of the best computer of Frontier

yearsToSec'' :: Double

Seconds in a year

frontierDouble :: Double

Power of operation of the Frontier since the beginning of time | >>> frontierDouble 5.158595808e35

B)

time :: Integer

totalYears :: Integer

>>> totalYears
317

C)

daysToSec :: Integer

Seconds in a day

hoursToSec :: Integer

Seconds in an hour

minsToSec :: Integer

Segundos totales en un minuto

years :: Integer

Caclulo de los años equivalentes a time

days :: Integer

Calculo de las horas equivalentes a time quitando years

hours :: Integer

Calculo de las horas equivalentes a time quitando days y years

mins :: Integer

Calculo de las horas equivalentes a time quitando days, years y hours

segs :: Integer

Calculo de las horas equivalentes a time quitando days, years, hours y mins

solc :: (Integer, Integer, Integer, Integer, Integer)

Tupla con el tiempo total de time en años, dias, horas, minutos y segundos

D)

sold :: Integer -> (Integer, Integer, Integer, Integer, Integer)

Descripción

Coge el tiempo dado y lo convierte a (Años, Días, Horas, Minutos, Segundos)

Ejemplos

>>> sold 1000000
(0,11,13,46,40)

EJERCICIO 2

f :: Int -> Int -> Int

Descripción

Función f usando notación infija

Ejemplos

>>> f 7 4
-14

f' :: Int -> Int -> Int

Descripción

Función f usando notación prefija

Ejemplo

>>> f' 7 4
-14

g :: Int -> Int

Descripción

Función g

Ejemplo

>>> g (-2)
32

h :: Int -> Int -> Int -> Int

Descripción

Función h usando notación infija

Ejemplo

>>> h 1 2 3
0

h' :: Int -> Int -> Int -> Int

Descripción

Función h usando notación posfija

Ejemplo

>>> h' 1 2 3
0

i :: Int -> Int -> Int

Descripción

Función i usando notación guardas

Ejemplo

>>> i 1 2
0

i' :: Int -> Int -> Int

Descripción

Función i usando notación if _ then _ else

Ejemplo

>>> i 1 2
0

EJERCICIO 3

A)

digitos :: Int -> Int

B)

sumaDigitos :: Int -> Int

reduccion :: Int -> Int

C)

perm :: Int -> Int

D)

var :: Int -> Int -> Int

E)

comb :: Int -> Int -> Int

EJERCICIO 4

factores :: Integral a => a -> [a]

Descripción

Calcula todos los divisores de un número dado

Ejemplos

>>> factores 10
[1,2,5,10]

isPrime :: Integral a => a -> Bool

Descripción

Comprueba si un número es primo

Usa factores

Ejemplos

>>> isPrime 10
False

>>> isPrime 5
True

EJERCICIO 5

fibonacci :: Int -> Integer

Descipción

Funcion de fibonacci de aridad 1

Ejemplos

>>> fibonacci 2
2

>>> fibonacci 100
573147844013817084101