Lote3

Description

EJERCICIO 1

data Complejo

Representación de un numero complejo

Constructors

C Float Float

EJERCICIO 2

data Arbol a

Constructors

N a [Arbol a]

showTree :: Show a => Int -> Int -> Arbol a -> [Char]

listaHojas :: Arbol a -> [a]

listaNodos :: Arbol a -> [a]

repMax :: Ord a => Arbol a -> Arbol a

EJERCICIO 3

class Medible a where

Clase para medir otros tipos de datos

Methods

medida

Arguments

:: a

-> Int
Coge el dato dado y saca una medida de él

EJERCICIO 4

data Exp

Constructors

Num Int

Mas Exp Exp

Menos Exp Exp

Por Exp Exp

Div Exp Exp

eval :: Exp -> Int

cuentaOps :: Exp -> Int

cambiaOps :: Exp -> Exp

operandos :: Exp -> [Int]

EJERCICIO 5

data Punto

Representación de un punto de forma (x,y)

Constructors

P
Constructor del punto

Fields

Int

Coordenada X

Int

Coordenada Y

pointSum :: (Ord a, Num a) => a -> a -> a

Descripción

Suma un numero a una parte de una coordenada dada. Tiene como rango [0 .. 100]

Ejemplos

>>> pointSum 100 10
9

>>> pointSum 0 (-1)
100

>>> pointSum 100 1
0

data Direccion

Representación de una dirección

Constructors

ARRIBA

ABAJO

IZQUIERDA

DERECHA

A)

mueve :: Punto -> Direccion -> Punto

Descripción

Mueve un punto en la dirección dada

Ejemplos

>>> mueve (P 0 1) IZQUIERDA
(100,1)

>>> mueve (P 0 1) DERECHA
(1,1)

>>> mueve (P 0 1) ARRIBA
(0,2)

>>> mueve (P 0 1) ABAJO
(0,0)

B)

destino :: Punto -> [Direccion] -> Punto

Descripción

Mueve un punto en las direcciones dadas

Ejemplos

>>> destino (P 0 0) [ARRIBA, ARRIBA, ABAJO, ABAJO, IZQUIERDA, DERECHA, IZQUIERDA, DERECHA]
(0,0)

>>> destino (P 0 0) [ARRIBA, ARRIBA, ABAJO, IZQUIERDA, IZQUIERDA, DERECHA]
(100,1)

C)

trayectoria :: Punto -> [Direccion] -> [Punto]

Descripción

Mueve un punto en las direcciones dadas mostrando los puntos por los que se han pasado

Ejemplos

>>> trayectoria (P 0 0) [ARRIBA, ARRIBA, ABAJO, ABAJO, IZQUIERDA, DERECHA, IZQUIERDA, DERECHA]
[(0,0),(0,1),(0,2),(0,1),(0,0),(100,0),(0,0),(100,0),(0,0)]

>>> trayectoria (P 0 0) [ARRIBA, ARRIBA, ABAJO, IZQUIERDA, IZQUIERDA, DERECHA]
[(0,0),(0,1),(0,2),(0,1),(100,1),(99,1),(100,1)]

inferior :: [Direccion] -> [Direccion] -> Bool

EJERCICIO 6

data Conjunto a

Constructors

Con Int [a]

toC :: Eq a => [a] -> Conjunto a

interseccion :: Eq a => Conjunto a -> Conjunto a -> Conjunto a

union :: Eq a => Conjunto a -> Conjunto a -> Conjunto a

mapSet :: (Eq a, Eq b) => (a -> b) -> Conjunto a -> Conjunto b

EJERCICIO 7

data Asec a b

Constructors

Void

Acons a (Asec b a)

long :: Num p => Asec a b -> p

nElem :: Int -> Asec a b -> Either a b

separa :: Asec a b -> ([a], [b])

showSec :: (Show a, Show b) => Asec a b -> String

EJERCICIO 8

palabras :: String -> IO Int

palabras' :: IO ()

getInt :: IO Int

promedia :: IO ()

promediaAux :: [Int] -> IO ()