Traductor

Description

Práctica final de Programación Declarativa (PD) de Ingeniería Infromática de la UCM

Types

data NodeElement

Nodo de la tabla de traduccion usando un hash con colisiones con valores [0 .. 9] para guardar valores

type Traduction = (String, String)

Alias para las tuplas usadas para representar traducciones

type Bucket = [Traduction]

Elementos de la tabla de traducción asignados a un valor concreto, se guardarán varias traducciones distintas

data Hash a

Árbol para almacenar todos los nodos, puede ser vacío o un nodo con ramas a ambos lados. Estas ramas también pueden ser vacias

Functions

Generators

leaf :: a -> Hash a

Description

Genera un árbol que solo tiene raiz, una hoja

Examples

Expand

>>> leaf (NE 1 [])
[1] -> []

genEmptyTree :: [Int] -> Hash NodeElement

Description

Genera un árbol binario de n elementos en una lista para meter las palabras dadas

Tree and List manipulators

insert :: Hash NodeElement -> Int -> Traduction -> Hash NodeElement

Description

Inserta un elemento en el indice dado

Examples

Expand

>>> insert (Node (leaf (NE 0 [])) (NE 1 []) Nil) 1 ("a","b")
[0] -> []
[1] -> [("a","b")]

>>> insert Nil 1 ("a","b")
[1] -> [("a","b")]

insertList :: [Traduction] -> Traduction -> [Traduction]

Description

Inserción ordenada de traducciones en una lista.

inorder :: Show a => Hash a -> String

Description

Muestra el inorden del árbol dado

busqueda :: Hash NodeElement -> Int -> String -> String

Description

Busca una palabra en el listado dentro del árbol dado.

Usa el índice para encontrar el nodo del listado

Examples

Expand

>>> busqueda (insert genEmptyTree (hash "hola") ("hola", "hello")) (hash "hola") "hola"
"hello"

>>> busqueda (genEmptyTree) (hash "hola") "hola"
"*** Exception: Palabra no encontrada
CallStack (from HasCallStack):
  error, called at hash.hs:190:28 in main:Traductor

busquedaLista :: [Traduction] -> String -> String

Description

Busqueda de una palabra en un listado de traducciones para conseguir la traducción.

Examples

Expand

>>> busquedaLista [] "hola"
"Palabra no encontrada"

>>> busquedaLista [("Hola", "Hello"), ("A", "B")] "Hola"
"Hello"

evens :: [a] -> [a]

Description

Función auxiliar de odds para eliminar los indices pares de una lista

odds :: [a] -> [a]

Description

Elimina los indices pares de una lista

Utilities

hash :: String -> Int

Description

Hace un hash de una palabra y se queda con el último digito para usarlo de índice

Examples

Expand

>>> hash "Hola"
8

>>> hash "Adios"
0

calcMeanWordCount :: String -> Float

Description

Caclula la media de letras por palabra en una frase.

Usa foldl para recorrer cada una de las palabras y añadir sus letras al numero total

Mas tarde se divide entre el numero de palabras para sacar la media de letras sin contar los espacios en blanco

Examples

Expand

>>> calcMeanWordCount "a b c d"
1

>>> calcMeanWordCount "ABCD"
4

calcMeanWordCountExample :: String -> IO ()

Description

Coje el texto del fichero dado a la función y calcula la media de letras por palabra.

Examples

Expand

>>> calcMeanWordCountExample "./pruebas/quijote.txt"
4.5021253

>>> calcMeanWordCountExample "./pruebas/call_of_cthulhu.txt"
4.8737144

rellenadoPorFichero :: IO (Hash NodeElement)

Description

Pide un fichero para generar el árbol de traducciones

rellenadoPorFicheroFijo :: IO (Hash NodeElement)

Description

Lectura del fichero "datos.txt" para cargar un arbol de traducciones

IO

lecturaPalabras :: String -> IO String

Description

Lee palabras por el teclado y devuelve un String con todas las palabras introducidas

mostrarTraducciones :: IO ()

Description

Traduce las palabras dadas por el usuario por teclado.

Genera un árbol vacío, lo llena con un archivo de traducciones y lo usa para traducir una serie de palabras dadas

demoMode :: IO ()

Description

Modo para probar todas las funciones implementadas en el módulo