Lote1

Description

EJERCICIO 1

A)

petaFlop :: Integer

Flops in a petaFlop

fornl :: Integer

Computing power of the best computer of Frontier

yearsToSec :: Integer

Seconds in a year

frontierInteger :: Integer

Power of operation of the Frontier since the beginning of time | >>> frontierInteger 515859580800000000000000000000000000

petaFlop' :: Int

Flops in a petaFlop

fornl' :: Int

Computing power of the best computer of Frontier

yearsToSec' :: Int

Seconds in a year

frontierInt :: Int

Power of operation of the Frontier since the beginning of time | >>> frontierInt -7928755680649936896

petaFlop'' :: Double

Flops in a petaFlop

fornl'' :: Double

Computing power of the best computer of Frontier

yearsToSec'' :: Double

Seconds in a year

frontierDouble :: Double

Power of operation of the Frontier since the beginning of time | >>> frontierDouble 5.158595808e35

B)

time :: Integer

Time for the exercise

totalYears :: Integer

>>> totalYears
317

C)

daysToSec :: Integer

Seconds in a day

hoursToSec :: Integer

Seconds in an hour

minsToSec :: Integer

Segundos totales en un minuto

years :: Integer

Caclulo de los años equivalentes a time

days :: Integer

Calculo de las horas equivalentes a time quitando years

hours :: Integer

Calculo de las horas equivalentes a time quitando days y years

mins :: Integer

Calculo de las horas equivalentes a time quitando days, years y hours

segs :: Integer

Calculo de las horas equivalentes a time quitando days, years, hours y mins

solc :: (Integer, Integer, Integer, Integer, Integer)

Tupla con el tiempo total de time en años, dias, horas, minutos y segundos

D)

sold :: Integer -> (Integer, Integer, Integer, Integer, Integer)

Descripción

Coge el tiempo dado y lo convierte a (Años, Días, Horas, Minutos, Segundos)

Ejemplos

>>> sold 1000000
(0,11,13,46,40)

EJERCICIO 2

f :: Int -> Int -> Int

Descripción

Función f usando notación infija

Ejemplos

>>> f 7 4
-14

f' :: Int -> Int -> Int

Descripción

Función f usando notación prefija

Ejemplo

>>> f' 7 4
-14

g :: Int -> Int

Descripción

Función g

Ejemplo

>>> g (-2)
32

h :: Int -> Int -> Int -> Int

Descripción

Función h usando notación infija

Ejemplo

>>> h 1 2 3
0

h' :: Int -> Int -> Int -> Int

Descripción

Función h usando notación posfija

Ejemplo

>>> h' 1 2 3
0

i :: Int -> Int -> Int

Descripción

Función i usando notación guardas

Ejemplo

>>> i 1 2
0

i' :: Int -> Int -> Int

Descripción

Función i usando notación if _ then _ else

Ejemplo

>>> i 1 2
0

EJERCICIO 3

A)

digitos :: Int -> Int

Description

Gets the number of digits of a given value

Example

>>> digitos 1
1

>>> digitos 123912891289
12

B)

sumaDigitos :: Int -> Int

Description

Adds every single digit of a given number

Example

>>> sumaDigitos 1
1

>>> sumaDigitos 18982918192283429138194213189
82

reduccion :: Int -> Int

Description

Adds digits (sumaDigitos) from a number until de result is one single digit. Single digits bellow 0 are converted to positive values

Example

>>> reduccion (-1)
1

>>> reduccion 12893213913819238192312819
2

C)

perm :: Integer -> Integer

Description

Permutations of a given number

Example

>>> perm 8
40320

D)

var :: Integer -> Integer -> Integer

Description

Variations of two given numbers

Example

>>> var 2 1
2

>>> var 123 5
25928567280

E)

comb :: Integer -> Integer -> Integer

Description

Combinations of two given numbers

Examples

>>> comb 2 1
2

>>> comb 12425 12423
77184100

EJERCICIO 4

factores :: Integral a => a -> [a]

Descripción

Calcula todos los divisores de un número dado

Ejemplos

>>> factores 10
[1,2,5,10]

isPrime :: Integral a => a -> Bool

Descripción

Comprueba si un número es primo

Usa factores

Ejemplos

>>> isPrime 10
False

>>> isPrime 5
True

EJERCICIO 5

fibonacci :: Int -> Integer

Descipción

Funcion de fibonacci de aridad 1

Ejemplos

>>> fibonacci 2
2

>>> fibonacci 100
573147844013817084101

EJERCICIO 6

y :: Bool -> Bool -> Bool

Conjunction operator

o :: Bool -> Bool -> Bool

Disjunction operator

no :: Bool -> Bool

Not operator

implica :: Bool -> Bool -> Bool

Implication operator

eq :: Bool -> Bool -> Bool

Equivalence operator

oEx :: Bool -> Bool -> Bool

Exclusive disjunction operator

(&&&) :: Bool -> Bool -> Bool infixr 3

Conjunction operator

(|||) :: Bool -> Bool -> Bool infix 3

Disjunction operator

(-->) :: Bool -> Bool -> Bool infix 3

Implication operator

(===) :: Bool -> Bool -> Bool infix 3

Equivalence operator

(||=) :: Bool -> Bool -> Bool infix 3

Exclusive disjunction operator

EJERCICIO 8

f8 :: Int -> Int -> Int -> Int

Description

La función toma tres argumentos y cumple con las siguientes condiciones:

(i) Ser estricta en el primer argumento.

(ii) No ser estricta ni en el segundo ni en el tercer argumento.

(iii) Ser conjuntamente estricta en el segundo y tercer argumento.

Examples

>>> f8 1 undefined undefined
1

>>> f8 0 1 2
3

>>> f8 0 1 undefined
Prelude.undefined

EJERCICIO 9

media :: (Fractional a, Integral a, Foldable t) => t a -> a

Description

Arithmetic mean of a list of numbers

Example

>>> media [1,2,3,4]
2.5

EJERCICIO 10

last' :: [a] -> a

Description

Returns the last element of a given list

Example

>>> last' [1,2,3,4]
4

init' :: [a] -> [a]

Description

Returns all element except the las one of a list

Examples

>>> init [1,2,3,4]
[1,2,3]

initLast :: [a] -> ([a], a)

Description

Returns a tuple with the init' of a list and the last' of a list

Example

>>> initLast [1,2,3,4]
([1,2,3],4)

concat' :: [[a]] -> [a]

Description

Concatenates all given lists into one

Example

>>> concat' [[1],[2,3],[4],[5]]
[1,2,3,4,5]

auxList :: Int -> Int -> Int -> [a] -> [a]

Description

Returns the elements of a list between two given indexes

Examples

>>> auxList 0 0 2 [1,2,3,4,5]
[1,2,3]

take' :: Int -> [a] -> [a]

Description

Takes the n first elements of a list

Example

>>> take' 2 [1,2,3,4]
[1,2]

drop' :: Int -> [a] -> [a]

Description

Erases the n first elements of a list

Example

>>> drop' 2 [1,2,3,4]
[3,4]

splitAt' :: Int -> [a] -> ([a], [a])

Description

Splits a list in the given index

Example

>>> splitAt' 2 [1,2,3,4]
([1,2],[3,4])

nub :: Eq a => [a] -> [a]

Description

Returns a list without the repeated values of the given list

Example

>>> nub [1,2,3,4,1,2,3,4,3,2,6,4,2,0]
[1,2,3,4,6,0]

quicksort :: Ord a => [a] -> [a]

Description

Quicksort implementation

Examples

>>> quicksort [4,7,3,6,1,89,52,76,13,0,23]
[0,1,3,4,6,7,13,23,52,76,89]

and' :: [Bool] -> Bool

Description

Conjunction of al elements of a list

Example

>>> and' [True,False, True ]
False

or' :: [Bool] -> Bool

Description

Disjunction of al elements of a list

Example

>>> or' [True,False, True ]
True

sum' :: [Int] -> Int

Description

Sum of al the numbers in the list

Example

>>> sum' [1..100]
5050

product' :: [Integer] -> Integer

Description

Product of al the numbers in the list

Example

>>> product' [2..5]
120

lmedia :: [[[a]]] -> Float

Description

Mean of sizes of al elements in a list of lists

EJERCICIO 11

A)

reverse' :: [a] -> [a]

Description

Reverse a given list in o(n^2) time

Example

>>> reverse' [1,2,3,4]
[4,3,2,1]

B)

reverse'' :: [a] -> [a]

Description

Reverse a given list in o(n) time

Example

>>> reverse'' [1,2,3,4]
[4,3,2,1]